+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to `bonjour-hap` will be documented in this file. This proje
 - fix(types): align `index.d.ts` with runtime
 - chore: dependency updates
 - fix(server): continue past unanswered questions in multi-question queries
+- fix(browser): handle null opts in constructor
 
 ## v3.10.2 (2026-05-04)